PackStream refactoring
