+++ driver/src/main/java/org/neo4j/driver/packstream/EndOfStream.java
@@ -16,29 +16,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.neo4j.driver.packstream;
 
-
-package org.neo4j.driver.v1.hydration;
-
-public class StructureHeader
+public class EndOfStream extends PackStreamException
 {
-    private final int size;
-    private final byte signature;
+    private static final long serialVersionUID = 5102836237108105603L;
 
-    public StructureHeader(int size, byte signature)
+    public EndOfStream(String message)
     {
-        this.size = size;
-        this.signature = signature;
+        super(message);
     }
-
-    public int size()
-    {
-        return size;
-    }
-
-    public byte signature()
-    {
-        return signature;
-    }
-
 }
