+++ driver/src/main/java/org/neo4j/driver/packstream/StructureHeader.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright (c) 2002-2017 "Neo Technology,"
+ * Network Engine for Objects in Lund AB [http://neotechnology.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.neo4j.driver.packstream;
+
+public class StructureHeader
+{
+    private final int size;
+    private final byte signature;
+
+    public StructureHeader(int size, byte signature)
+    {
+        this.size = size;
+        this.signature = signature;
+    }
+
+    public int size()
+    {
+        return size;
+    }
+
+    public byte signature()
+    {
+        return signature;
+    }
+
+}
