+++ driver/src/main/java/org/neo4j/driver/packstream/io/BufferedChannelOutput.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.neo4j.driver.internal.packstream;
+package org.neo4j.driver.packstream.io;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
