+++ driver/src/main/java/org/neo4j/driver/packstream/io/PackOutput.java
@@ -16,7 +16,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.neo4j.driver.internal.packstream;
+package org.neo4j.driver.packstream.io;
+
+import org.neo4j.driver.packstream.PackStream;
 
 import java.io.IOException;
