+++ driver/src/main/java/org/neo4j/driver/v1/types/GraphValueAdapter.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.neo4j.driver.v1.hydration;
+package org.neo4j.driver.v1.types;
 
 import org.neo4j.driver.internal.value.ValueAdapter;
