+++ driver/src/main/java/org/neo4j/driver/v1/types/SelfContainedNode.java
@@ -16,13 +16,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.neo4j.driver.v1.hydration;
+package org.neo4j.driver.v1.types;
 
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Map;
 
-import org.neo4j.driver.v1.types.Node;
 import org.neo4j.driver.v1.Value;
 
 /**
