+++ driver/src/test/java/org/neo4j/driver/internal/InternalMapAccessorWithDefaultValueTest.java
@@ -31,10 +31,10 @@ import org.neo4j.driver.internal.value.FloatValue;
 import org.neo4j.driver.internal.value.IntegerValue;
 import org.neo4j.driver.internal.value.ListValue;
 import org.neo4j.driver.internal.value.MapValue;
-import org.neo4j.driver.v1.hydration.NodeValue;
+import org.neo4j.driver.v1.types.NodeValue;
 import org.neo4j.driver.internal.value.NullValue;
-import org.neo4j.driver.v1.hydration.PathValue;
-import org.neo4j.driver.v1.hydration.RelationshipValue;
+import org.neo4j.driver.v1.types.PathValue;
+import org.neo4j.driver.v1.types.RelationshipValue;
 import org.neo4j.driver.internal.value.StringValue;
 import org.neo4j.driver.v1.Record;
 import org.neo4j.driver.v1.Value;
@@ -43,9 +43,9 @@ import org.neo4j.driver.v1.types.Entity;
 import org.neo4j.driver.v1.types.Node;
 import org.neo4j.driver.v1.types.Path;
 import org.neo4j.driver.v1.types.Relationship;
-import org.neo4j.driver.v1.hydration.SelfContainedNode;
-import org.neo4j.driver.v1.hydration.SelfContainedPath;
-import org.neo4j.driver.v1.hydration.SelfContainedRelationship;
+import org.neo4j.driver.v1.types.SelfContainedNode;
+import org.neo4j.driver.v1.types.SelfContainedPath;
+import org.neo4j.driver.v1.types.SelfContainedRelationship;
 
 import static java.util.Arrays.asList;
 import static org.hamcrest.CoreMatchers.equalTo;
