+++ driver/src/test/java/org/neo4j/driver/v1/tck/tck/util/TestRelationshipValue.java
@@ -20,8 +20,8 @@ package org.neo4j.driver.v1.tck.tck.util;
 
 import java.util.Map;
 
-import org.neo4j.driver.v1.hydration.SelfContainedRelationship;
-import org.neo4j.driver.v1.hydration.RelationshipValue;
+import org.neo4j.driver.v1.types.SelfContainedRelationship;
+import org.neo4j.driver.v1.types.RelationshipValue;
 import org.neo4j.driver.v1.types.Entity;
 import org.neo4j.driver.v1.types.Relationship;
 import org.neo4j.driver.v1.Value;
